feat: isolate `DefId`: faster build times
