+++ frontend/exporter/src/types/replaced.rs
@@ -2,7 +2,6 @@ use crate::prelude::*;
 
 pub type Path = Vec<String>;
 
-pub type Symbol = String;
 impl<'t, S> SInto<S, Symbol> for rustc_span::symbol::Symbol {
     fn sinto(&self, _s: &S) -> Symbol {
         self.to_ident_string()
